// src/app/shared/enums/bic.creation.type.ts
export enum OpenQuestionOptionsOrder {
  Order,
  Random,
  Alphabetic,
}

export enum SectionTitleType {
  First = 'Section 1',
  Second = 'Section 2',
  Third = 'Section 3',
  Fourth = 'Section 4',
  Fifth = 'Section 5',
  Sixth = 'Section 6',
  Seventh = 'Section 7',
  Eighth = 'Section 8',
}

export enum IncorrectSection {
  NoErrors = 0,
  First = 1,
  Second = 2,
  Third = 3,
  Fourth = 4,
  Fifth = 5,
  Sixth = 6,
  Seventh = 7,
}